organizacion de logicas en componentes
